move pg controller over to other controllers
